+++ core/src/model/battlefield/army/ArmyManager.java
@@ -18,14 +18,14 @@ import model.battlefield.army.effects.PersistentEffect;
  * @author Benoît
  */
 public class ArmyManager {
-	private static Map<Long, Unit> units = new HashMap<Long, Unit>();
+	private static Map<Integer, Unit> units = new HashMap<Integer, Unit>();
 	private static List<PersistentEffect> persistenteffects = new ArrayList<>();
 	private static List<Projectile> projectiles = new ArrayList<>();
 
 	public static void update(double elapsedTime) {
-		Iterator<Entry<Long, Unit>> unitIterator = units.entrySet().iterator();
+		Iterator<Entry<Integer, Unit>> unitIterator = units.entrySet().iterator();
 		while (unitIterator.hasNext()) {
-			Entry<Long, Unit> entry = unitIterator.next();
+			Entry<Integer, Unit> entry = unitIterator.next();
 			Unit u = entry.getValue();
 			if (u.destroyed()) {
 				u.faction.getUnits().remove(u);
